rename package name for tests
